Updated the graphite template to work in both Ubuntu 12.04 and 14.04.

+++ templates/graphite/config/rubber/deploy-graphite.rb
@@ -22,33 +22,35 @@ namespace :rubber do
     end
 
     task :install_graphite_from_repo, :roles => [:graphite_server, :graphite_web] do
-      rubber.sudo_script 'install_graphite_from_repo', <<-ENDSCRIPT
-        if [[ ! -d "/opt/graphite" ]]; then
-          mkdir -p /tmp/graphite_install
-          cd /tmp/graphite_install
-          bzr branch lp:graphite -r #{rubber_env.graphite_repo_version}
+      if old_ubuntu?(capture('hostname').chomp)
+        rubber.sudo_script 'install_graphite_from_repo', <<-ENDSCRIPT
+          if [[ ! -d "/opt/graphite" ]]; then
+            mkdir -p /tmp/graphite_install
+            cd /tmp/graphite_install
+            bzr branch lp:graphite -r #{rubber_env.graphite_repo_version}
 
-          pushd graphite/whisper
-          python setup.py install
-          popd
+            pushd graphite/whisper
+            python setup.py install
+            popd
 
-          pushd graphite/carbon
-          python setup.py install
-          popd
+            pushd graphite/carbon
+            python setup.py install
+            popd
 
-          pushd graphite
-          python setup.py install
-          popd
-        fi
-      ENDSCRIPT
+            pushd graphite
+            python setup.py install
+            popd
+          fi
+        ENDSCRIPT
+      end
     end
 
     desc <<-DESC
-      Cleans out old whisper storage files for non-existant instances so they don't show in webui
+      Cleans out old whisper storage files for non-existent instances so they don't show in webui
     DESC
     task :clean_storage, :roles => [:graphite_server] do
       active_instances = rubber_instances.collect {|ic| ic.name }.sort
-      stored_instances = capture("ls /opt/graphite/storage/whisper/servers/").split.sort
+      stored_instances = capture("ls #{rubber_env.graphite_storage_dir}/whisper/servers/").split.sort
       cleaning_instances = stored_instances - active_instances
 
       if cleaning_instances.size > 0
@@ -61,7 +63,7 @@ namespace :rubber do
         end
 
         if do_clean
-          rsudo "rm -rf /opt/graphite/storage/whisper/servers/{#{cleaning_instances.join(',')}}"
+          rsudo "rm -rf #{rubber_env.graphite_storage_dir}/whisper/servers/{#{cleaning_instances.join(',')}}"
         end
 
       end
@@ -72,45 +74,49 @@ namespace :rubber do
       rubber.allow_optional_tasks(self)
 
       if Rubber::Configuration.rubber_env.graphite_repo_version
-        after "rubber:install_packages", "rubber:graphite:install_graphite_from_repo"
+        after "rubber:install_gems", "rubber:graphite:install_graphite_from_repo"
       else
-        after "rubber:install_packages", "rubber:graphite:server:install"
+        after "rubber:install_gems", "rubber:graphite:server:install"
       end
 
-      after "rubber:bootstrap", "rubber:graphite:server:bootstrap"
+      after "rubber:graphite:server:install", "rubber:graphite:server:bootstrap"
 
       desc <<-DESC
         Installs graphite server components
       DESC
       task :install, :roles => :graphite_server do
-        rubber.sudo_script 'install_graphite_server', <<-ENDSCRIPT
-          if [[ ! -f "/opt/graphite/bin/carbon-cache.py" ]]; then
-            wget --content-disposition -qNP /tmp #{rubber_env.graphite_whisper_package_url}
-            tar -C /tmp -zxf /tmp/whisper-#{rubber_env.graphite_version}.tar.gz
-            cd /tmp/whisper-#{rubber_env.graphite_version}
-            python setup.py install
-            cd /tmp
-            rm -rf whisper-#{rubber_env.graphite_version}
-            rm whisper-#{rubber_env.graphite_version}.tar.gz
+        if old_ubuntu?(capture('hostname').chomp)
+          rubber.sudo_script 'install_graphite_server', <<-ENDSCRIPT
+            if [[ ! -f "/opt/graphite/bin/carbon-cache.py" ]]; then
+              wget --content-disposition -qNP /tmp #{rubber_env.graphite_whisper_package_url}
+              tar -C /tmp -zxf /tmp/whisper-#{rubber_env.graphite_version}.tar.gz
+              cd /tmp/whisper-#{rubber_env.graphite_version}
+              python setup.py install
+              cd /tmp
+              rm -rf whisper-#{rubber_env.graphite_version}
+              rm whisper-#{rubber_env.graphite_version}.tar.gz
 
-            wget --content-disposition -qNP /tmp #{rubber_env.graphite_carbon_package_url}
-            tar -C /tmp -zxf /tmp/carbon-#{rubber_env.graphite_version}.tar.gz
-            cd /tmp/carbon-#{rubber_env.graphite_version}
-            python setup.py install
-            cd /tmp
-            rm -r carbon-#{rubber_env.graphite_version}
-            rm carbon-#{rubber_env.graphite_version}.tar.gz
+              wget --content-disposition -qNP /tmp #{rubber_env.graphite_carbon_package_url}
+              tar -C /tmp -zxf /tmp/carbon-#{rubber_env.graphite_version}.tar.gz
+              cd /tmp/carbon-#{rubber_env.graphite_version}
+              python setup.py install
+              cd /tmp
+              rm -r carbon-#{rubber_env.graphite_version}
+              rm carbon-#{rubber_env.graphite_version}.tar.gz
 
-            rm -rf /opt/graphite/storage
-            mkdir #{rubber_env.graphite_storage_dir}
-            chown www-data:www-data #{rubber_env.graphite_storage_dir}
-            ln -s #{rubber_env.graphite_storage_dir} /opt/graphite/storage
-          fi
-        ENDSCRIPT
+              rm -rf #{rubber_env.graphite_storage_dir}
+              mkdir #{rubber_env.graphite_storage_dir}
+              chown www-data:www-data #{rubber_env.graphite_storage_dir}
+              ln -s #{rubber_env.graphite_storage_dir} /opt/graphite/storage
+            fi
+          ENDSCRIPT
+        end
+
+        create_storage_directory
       end
 
       task :bootstrap, :roles => :graphite_server do
-        exists = capture("echo $(ls /opt/graphite/storage/whisper/ 2> /dev/null)")
+        exists = capture("echo $(ls #{rubber_env.graphite_storage_dir}/whisper/ 2> /dev/null)")
         if exists.strip.size == 0
           rubber.update_code_for_bootstrap
 
@@ -122,12 +128,20 @@ namespace :rubber do
 
       desc "Start graphite system monitoring"
       task :start, :roles => :graphite_server do
-        rsudo "service graphite-server start"
+        if old_ubuntu?(capture('hostname').chomp)
+          rsudo 'service graphite-server start'
+        else
+          rsudo 'service carbon-cache start'
+        end
       end
 
       desc "Stop graphite system monitoring"
       task :stop, :roles => :graphite_server do
-        rsudo "service graphite-server stop || true"
+        if old_ubuntu?(capture('hostname').chomp)
+          rsudo 'service graphite-server stop || true'
+        else
+          rsudo 'service carbon-cache stop || true'
+        end
       end
 
       desc "Restart graphite system monitoring"
@@ -155,28 +169,32 @@ namespace :rubber do
         after "rubber:graphite:server:install", "rubber:graphite:web:install"
       end
 
-      after "rubber:graphite:server:bootstrap", "rubber:graphite:web:bootstrap"
+      after "rubber:graphite:web:install", "rubber:graphite:web:bootstrap"
 
       desc <<-DESC
         Installs graphite web components
       DESC
       task :install, :roles => :graphite_web do
-        rubber.sudo_script 'install_graphite_web', <<-ENDSCRIPT
-          if [[ ! -d "/opt/graphite/webapp" ]]; then
-            wget --content-disposition -qNP /tmp #{rubber_env.graphite_web_package_url}
-            tar -C /tmp -zxf /tmp/graphite-web-#{rubber_env.graphite_version}.tar.gz
-            cd /tmp/graphite-web-#{rubber_env.graphite_version}
-            # python check-dependencies.py
-            python setup.py install
-            cd /tmp
-            rm -r graphite-web-#{rubber_env.graphite_version}
-            rm graphite-web-#{rubber_env.graphite_version}.tar.gz
-          fi
-        ENDSCRIPT
+        if old_ubuntu?(capture('hostname').chomp)
+          rubber.sudo_script 'install_graphite_web', <<-ENDSCRIPT
+            if [[ ! -d "/opt/graphite/webapp" ]]; then
+              wget --content-disposition -qNP /tmp #{rubber_env.graphite_web_package_url}
+              tar -C /tmp -zxf /tmp/graphite-web-#{rubber_env.graphite_version}.tar.gz
+              cd /tmp/graphite-web-#{rubber_env.graphite_version}
+
+              # python check-dependencies.py
+              python setup.py install
+
+              cd /tmp
+              rm -r graphite-web-#{rubber_env.graphite_version}
+              rm graphite-web-#{rubber_env.graphite_version}.tar.gz
+            fi
+          ENDSCRIPT
+        end
       end
 
       task :bootstrap, :roles => :graphite_web do
-        exists = capture("echo $(ls /opt/graphite/storage/graphite.db 2> /dev/null)")
+        exists = capture("echo $(ls #{rubber_env.graphite_storage_dir}/graphite.db 2> /dev/null)")
         if exists.strip.size == 0
           rubber.update_code_for_bootstrap
 
@@ -188,19 +206,42 @@ namespace :rubber do
           env = rubber_cfg.environment.bind("graphite_web", web_instance.name)
           email = env.admin_email
 
-          rubber.sudo_script 'bootstrap_graphite_web', <<-ENDSCRIPT
-            cd /opt/graphite/webapp/graphite
-            python manage.py syncdb --noinput
-            python manage.py createsuperuser --username admin --email #{email} --noinput
-            python manage.py shell <<EOF
+          create_storage_directory
+
+          if old_ubuntu?(capture('hostname').chomp)
+            rubber.sudo_script 'bootstrap_graphite_web', <<-ENDSCRIPT
+              mkdir -p #{rubber_env.graphite_storage_dir}/log/webapp/
+              chown -R www-data:www-data #{rubber_env.graphite_storage_dir}/log/
+
+              cd /opt/graphite/webapp/graphite
+              python manage.py syncdb --noinput
+              python manage.py createsuperuser --username admin --email #{email} --noinput
+              python manage.py shell <<EOF
 from django.contrib.auth.models import User
 u = User.objects.get(username__exact='admin')
 u.set_password('admin1')
 u.save()
 EOF
-            mkdir -p /opt/graphite/storage/log/webapp/
-            chown -R www-data:www-data /opt/graphite/storage/
-          ENDSCRIPT
+            ENDSCRIPT
+          else
+            rubber.sudo_script 'bootstrap_graphite_web', <<-ENDSCRIPT
+              # Ubuntu 14.04 ships with a broken graphite-web package.  It renames the build-index.sh file to be a binary
+              # on the PATH, but it fails to update any of the code that references the original file location, causing
+              # graphite-web to fail on initial load.  We fix that here by setting up a symlink to the renamed binary.
+
+              mkdir -p /usr/share/graphite-web/bin/
+              ln -s /usr/bin/graphite-build-search-index /usr/share/graphite-web/bin/build-index.sh
+
+              graphite-manage syncdb --noinput
+              graphite-manage createsuperuser --username admin --email #{email} --noinput
+              graphite-manage shell <<EOF
+from django.contrib.auth.models import User
+u = User.objects.get(username__exact='admin')
+u.set_password('admin1')
+u.save()
+EOF
+            ENDSCRIPT
+          end
 
           restart
         end
@@ -229,6 +270,21 @@ EOF
         rsudo "netstat -tupln | grep uwsgi || true"
       end
 
+    end
+
+    def create_storage_directory
+      owner = old_ubuntu?(capture('hostname').chomp) ? 'www-data' : '_graphite'
+
+      rubber.sudo_script 'create_graphite_storage_directory', <<-ENDSCRIPT
+        if [[ ! -e #{rubber_env.graphite_storage_dir} ]]; then
+          mkdir -p #{rubber_env.graphite_storage_dir}
+          chown -R #{owner}:www-data #{rubber_env.graphite_storage_dir}
+        fi
+      ENDSCRIPT
+    end
+
+    def old_ubuntu?(host)
+      %w[10.04 12.04].include?(rubber_instances[host].os_version)
     end
 
   end
